% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_process_website.R
\name{generate_process_website}
\alias{generate_process_website}
\title{Generate website for selected processes}
\usage{
generate_process_website(process_files = list.files("inst/processes",
  full.names = TRUE), website = "inst/diagramTextButton.html",
  write_to = "diagram_out.html")
}
\arguments{
\item{process_files}{vector of files names with process definitions}

\item{website}{path to website scheme}

\item{write_to}{path to html file in which the final website will be written}
}
\description{
Generate website for selected processes
}
