% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_process.R
\name{read_process}
\alias{read_process}
\title{Read process from a config file}
\usage{
read_process(file = "processes/general.txt")
}
\arguments{
\item{file}{name of a file with title, phases, steps, and efforts}
}
\value{
list with config
}
\description{
Read process from a config file
}
\examples{
config <- read_process()
str(config)
}
