% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_process.R
\name{read_process}
\alias{read_process}
\title{Read process from a config file}
\usage{
read_process(file = "inst/processes/0general.txt")
}
\arguments{
\item{file}{name of a file with title, phases, steps, and efforts}
}
\value{
list with process detailes
}
\description{
Read process from a config file
}
\examples{
\dontrun{
  process <- read_process("inst/processes/0general.txt")
  str(process)
}
}
